tech: Move UrlSlicer in server where it is used
