+++ server/src/main/java/com/chutneytesting/agent/domain/explore/ExploreAgentsService.java
@@ -13,7 +13,6 @@ import com.chutneytesting.agent.domain.explore.ImmutableExploreResult.Link;
 import com.chutneytesting.agent.domain.network.NetworkDescription;
 import com.chutneytesting.engine.domain.delegation.ConnectionChecker;
 import com.chutneytesting.engine.domain.delegation.NamedHostAndPort;
-import com.chutneytesting.engine.domain.delegation.UrlSlicer;
 import com.chutneytesting.environment.api.dto.TargetDto;
 import com.google.common.collect.ImmutableSet;
 import java.util.Set;
